requirements.txt
anthropic
openai
requests
python-dotenv
python-telegram-bot
flask
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
authlib
razorpay
